requirements.txt
python==3.6.8
pandas==0.20.3
numpy==1.16.1
biopython==1.73
keras==2.1.4
plaidml.keras==0.5.0
tensorflow-gpu==1.3.0    
tensorflow-tensorboard==0.1.8
scikit-learn==0.20.2
scipy==1.2.1
matplotlib==3.0.3

